feat(authorisation) add some permissions

+++ routes/campgrounds.js
@@ -49,20 +49,40 @@ router.get('/:id', catchAsync(async (req, res) => {
 
 //This route will be used to show campground informaiton for editing
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
-    const campgrounds = await Campground.findById(req.params.id);
-    res.render('campgrounds/edit', { campgrounds })
+    const { id } = req.params;
+    const campgrounds = await Campground.findById(id)
+    if (!campgrounds) {
+        req.flash('error', 'Cannot find that campground!');
+        return res.redirect('/campgrounds');
+    }
+    if (!campgrounds.author.equals(req.user._id)) {
+        req.flash('error', 'You Do Not Have Permission to do that!');
+        return res.redirect(`/campgrounds/${campgrounds._id}`)
+    }
+    res.render('campgrounds/edit', { campgrounds });
 }));
 
 //this is the route for the edit form to push the update to the server
 router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You Do Not Have Permission to do that!');
+        return res.redirect(`/campgrounds/${campground._id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }));
 
 // Delete page for an individual campground
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You Do Not Have Permission to do that!');
+        return res.redirect(`/campgrounds/${campground._id}`)
+    }
     const campgrounds = await Campground.findByIdAndDelete(req.params.id);
     req.flash('success', 'Deleted the Campground');
     res.redirect('/campgrounds')
